cut huge src files
