+++ cgogn/core/CMakeLists.txt
@@ -27,8 +27,11 @@ target_sources(${PROJECT_NAME}
 		"${CMAKE_CURRENT_LIST_DIR}/types/cmap/cell.h"
 		"${CMAKE_CURRENT_LIST_DIR}/types/cmap/cmap_base.h"
 		"${CMAKE_CURRENT_LIST_DIR}/types/cmap/cmap_base_index.hpp"
-		"${CMAKE_CURRENT_LIST_DIR}/types/cmap/cmap_base_traversals.hpp"
-		"${CMAKE_CURRENT_LIST_DIR}/types/cmap/cmap_base_check.hpp"
+		"${CMAKE_CURRENT_LIST_DIR}/types/cmap/cmap_base_attribute.hpp"
+		"${CMAKE_CURRENT_LIST_DIR}/types/cmap/cmap_base_orbit_traversals.hpp"
+		"${CMAKE_CURRENT_LIST_DIR}/types/cmap/cmap_base_local_traversals.hpp"
+		"${CMAKE_CURRENT_LIST_DIR}/types/cmap/cmap_base_global_traversals.hpp"
+		"${CMAKE_CURRENT_LIST_DIR}/types/cmap/cmap_base_functions.hpp"
 		"${CMAKE_CURRENT_LIST_DIR}/types/cmap/cmap_base.cpp"
 		"${CMAKE_CURRENT_LIST_DIR}/types/cmap/cmap_info.h"
 		"${CMAKE_CURRENT_LIST_DIR}/types/cmap/cmap0.h"
