+++ cgogn/core/types/cmap/cmap_base_index.hpp
@@ -170,9 +170,6 @@ inline void init_cells_indexing(CMapBase& m, Orbit orbit)
 
 
 
-
-
-
 template <typename CELL, typename MESH>
 auto index_cells(MESH& m) -> std::enable_if_t<std::is_convertible_v<MESH&, CMapBase&>>
 {
@@ -199,110 +196,7 @@ auto index_cells(MESH& m) -> std::enable_if_t<std::is_convertible_v<MESH&, CMapB
 }
 
 
-
-template <typename T, typename CELL, typename MESH,
-		  typename std::enable_if_t<std::is_convertible_v<MESH&, CMapBase&>>* = nullptr>
-std::shared_ptr<typename mesh_traits<MESH>::template Attribute<T>> add_attribute(MESH& m, const std::string& name)
-{
-	static_assert(is_in_tuple<CELL, typename mesh_traits<MESH>::Cells>::value, "CELL not supported in this MESH");
-	if (!is_indexed<CELL>(m))
-		index_cells<CELL>(m);
-	CMapBase& mb = static_cast<CMapBase&>(m);
-	return mb.attribute_containers_[CELL::ORBIT].template add_attribute<T>(name);
-}
-
-
-
-template <typename T, typename CELL, typename MESH,
-		  typename std::enable_if_t<std::is_convertible_v<MESH&, CMapBase&>>* = nullptr>
-std::shared_ptr<CMapBase::Attribute<T>> get_attribute(const MESH& m, const std::string& name)
-{
-	static_assert(is_in_tuple<CELL, typename mesh_traits<MESH>::Cells>::value, "CELL not supported in this MESH");
-	return m.attribute_containers_[CELL::ORBIT].template get_attribute<T>(name);
-}
-
-
-template <typename CELL>
-void remove_attribute(CMapBase& m, const std::shared_ptr<CMapBase::AttributeGen>& attribute)
-{
-	m.attribute_containers_[CELL::ORBIT].remove_attribute(attribute);
-}
-
-
-template <typename CELL>
-void remove_attribute(CMapBase& m, CMapBase::AttributeGen* attribute)
-{
-	m.attribute_containers_[CELL::ORBIT].remove_attribute(attribute);
-}
-
-
-
-template <typename CELL, typename FUNC>
-void foreach_attribute(const CMapBase& m, const FUNC& f)
-{
-	using AttributeGen = CMapBase::AttributeGen;
-	static_assert(is_func_parameter_same<FUNC, const std::shared_ptr<AttributeGen>&>::value,
-				  "Wrong function attribute parameter type");
-	for (const std::shared_ptr<AttributeGen>& a : m.attribute_containers_[CELL::ORBIT])
-		f(a);
-}
-
-
-
-template <typename T, typename CELL, typename FUNC>
-void foreach_attribute(const CMapBase& m, const FUNC& f)
-{
-	using AttributeT = CMapBase::Attribute<T>;
-	using AttributeGen = CMapBase::AttributeGen;
-	static_assert(is_func_parameter_same<FUNC, const std::shared_ptr<AttributeT>&>::value,
-				  "Wrong function attribute parameter type");
-	for (const std::shared_ptr<AttributeGen>& a : m.attribute_containers_[CELL::ORBIT])
-	{
-		std::shared_ptr<AttributeT> at = std::dynamic_pointer_cast<AttributeT>(a);
-		if (at)
-			f(at);
-	}
-}
-
-
-
-inline std::shared_ptr<CMapBase::Attribute<Dart>> CMapBase::add_relation(const std::string& name)
-{
-	return relations_.emplace_back(darts_.add_attribute<Dart>(name));
-}
-
-
-
-
-
-
-template <typename T>
-T& get_attribute(CMapBase& m, const std::string& name)
-{
-	return m.get_attribute<T>(name);
-}
-
-template <typename CELL, typename MESH>
-auto get_mark_attribute(const MESH& m)
-	-> std::enable_if_t<std::is_convertible_v<MESH&, CMapBase&>, typename mesh_traits<MESH>::MarkAttribute*>
-{
-	static_assert(is_in_tuple<CELL, typename mesh_traits<MESH>::Cells>::value, "CELL not supported in this MESH");
-	if (!is_indexed<CELL>(m))
-		index_cells<CELL>(const_cast<MESH&>(m));
-	const CMapBase& mb = static_cast<const CMapBase&>(m);
-	return mb.attribute_containers_[CELL::ORBIT].get_mark_attribute();
-}
-
-template <typename CELL>
-void release_mark_attribute(const CMapBase& m, CMapBase::MarkAttribute* attribute)
-{
-	return m.attribute_containers_[CELL::ORBIT].release_mark_attribute(attribute);
-}
-
-
 } // namespace cgogn
 
-//#include <cgogn/core/utils/tuples.h>
 
-
-#endif // CGOGN_CORE_CMAP_CMAP_BASE_H_
+#endif // CGOGN_CORE_CMAP_CMAP_BASE_IND_H_
